requirements.txt
pystray~=0.19.5
pillow~=12.0.0
requests~=2.32.5